contextual logging
This moves the global logger into a separate file together with the functions
that implement contextual logging.

The advantage of SetLoggerWithOptions over alternatives like a
SetContextualLogger without options is that it can be used as a drop-in
replacement for SetLogger. This allows us to deprecate SetLogger. Whether we
actually do that is kept open for now.