fix for off-by-one stack unwinding
https://github.com/kubernetes/klog/pull/280/ turned println and printf into
wrapper functions, but the underlying *Depth functions kept passing the same 0
depth to l.output and thus any logr.Logger.

That off-by-one error broke JSON tests in Kubernetes, luckily well before any
of the PR went into a release. It would be nice to have regression testing for
this in klog, but we cannot depend on a Logger which does stack unwinding in
the code, so the only thing that could be done is some kind of GitHub action
which combines klog with some external logger.

The depth=1 in println is also not covered by unit tests, but seems to be
correct based on the JSON tests.