SetLogger/ClearLogger/SetFilter: document as not thread-safe
It has never been claimed that these functions are thread-safe and in practice
they weren't because read access was not protected by the same mutex as the
write access.

These functions were meant to be called during program initialization,
therefore the documentation is lacking, not the implementation.

Making them thread-safe would imply adding mutex locking to the implementation
of V, which most likely would have a noticeable effect.