log filter: ignored by V, used during log call
It's not entirely clear why the filter was copied into Verbose. The only
semantic difference is when Verbose get stored and then the filter gets
modified:

klogV := klog.V(5)
klog.SetLogFilter(...)
klogV.Info(...)

This seems like a very unlikely code pattern and goes against the intention of
first initializing logging, then using it.

Therefore we can undo the size increase in Verbose and only retrieve the filter
when it is needed.