klogr: report missing values with "(MISSING)"
klogr had a special value for a missing value in a key/value pair, so the
intention probably was to use that as replacement, the same way as klog does
it. But because of a bug in the code which removed duplicates, it treated a
missing value like nil. Now the behavior is consistent with klog.