test: move output test to example, include zapr
To reproduce the loss of verbosity in
klogr (https://github.com/kubernetes/klog/issues/294) we need a backend which
actually emits the verbosity. zapr does that and also happens to be the logging
implementation used in Kubernetes for JSON, so it makes sense to test with
that.

The expected output mapping may be useful also for testing in Kubernetes,
therefore it gets exported by k8s.io/klog/v2/test.

Because k8s.io/klog/v2 must not depend on zapr, the actual output test gets
moved to the examples. The downside is that "go test ./..." no longer runs it.
One has to remember to enter the "examples" directory. The GitHub action gets
updated to do that.