StartFlushDaemon: add API with more control over flushing
Previously, periodic flushing was started automatically when writing to
buffered files for the first time. When writing to a Logger which buffers
internally, then using the same periodic flushing makes sense. In that case it
has to be started explicitly.

The new call grants the caller control over the flush interval. Kubernetes has
a parameter for that.