promote non-test APIs to stable
As pointed out in
https://github.com/kubernetes/kubernetes/pull/108725#discussion_r830100796,
klog cannot carry "experimental" APIs that may end up in code that Kubernetes
depends on, whether it is directly or indirectly.

If that happens, Kubernetes might get stuck on a certain klog release because
one dependency depends on one klog release and another dependency on a
different, incompatible one.