hack/tools: drop dependency on golangci-lint
Depending on the more recent golangci-lint 1.43.0 also pulled in a newer mock
package, which caused problems when trying to use the logcheck plugin in
Kubernetes because there were problems with generated mock files when using the
newer mock.

golangci-lint isn't actually needed (anymore?) for building the plugin, so a
"go mod tidy" removed it.