use faster MergeKVs
Performance is better. This slightly changes the output of the traditional
klogr format:
- all key/value pairs gets sorted, not just each slice (an improvement?)
- a single space is printed where previously two were printed

klog no longer de-duplicates when the same key is passed twice in the same
call (WithValues, InfoS, etc.). That should not occur in practice and can be
prevented via code review and/or static code analysis, so we don't need to pay
the price of checking for it at runtime.