support logr.Marshaler
A type that implements only logr.Marshaler but not fmt.Stringer is useful for
delayed formatting. It is not required to return a struct. A plain string may
also be useful.

klog now supports such types by invoking MarshalLog once and then printing the
value as string (including multi-line support) if that is the result or with
the default "%+v".