ktesting: capture log data in memory
This is useful for tests that specifically want to check for expected log
output in the code that they are testing.

Capturing log data is done unconditionally. Having a configuration option for
it would be possible, but then usage and code become more complicated
for (probably) very little performance gains in tests that don't need the
captured data.