klog.Fatal: revert dumping of all stack backtraces to stderr
This is a revert of https://github.com/kubernetes/klog/pull/79.

Dumping the stack backtraces of all goroutines to stderr is not useful for
processes which have many goroutines. The console buffer overflows and the
original error which got dumped earlier is no longer visible. Even in CI
systems where stderr is captured the full dump is often not useful.

This was pointed out as a potential problem during the original PR review but
only got more attention after the updated klog got merged into Kubernetes and
developers there started to see the longer output.