github: run golangci-lint via action
The main advantage is that issues get posted as annotations, which makes them
easier to find when looking at a diff for a PR.

While at it, golangci-lint gets invoked so that it runs in its default
configuration plus the linters that were enabled explicitly before (misspell,
gofmt, revive as replacement for golint).

It also gets applied to the examples package.