add slog support
When compiled with Go >= 1.21, special support for log/slog gets enabled
automatically:

- The formatting of key/value pairs supports slog.Value and slog.LogValuer,
  regardless of where those come from.
- A slog.GroupValue is formatted as a JSON object.
- The logger returned by klog.Background() and the textlogger
  support usage as a backend for the slog API, via slogr.NewSlogHandler,
  with builtin support for handling a slog.Record. However, -vmodule
  does not work when called like that because stack unwinding during
  the Enabled check finds the wrong source code.
- KObj (= ObjectRef) and the type behind KObjSlice support the
  slog.LogValuer interface and thus get formatted properly by slog backends.

The klogr package doesn't get updated. It shouldn't be used anymore. To make
that clearer, it gets marked as "deprecated".

Examples demonstrate the resulting output. That output has to be exactly
the same when used as test, so pid and time get set to fixed values.