OWNERS: remove serathius, add mengjiao-liu, promote pohly
Serathius is focused on etcd these days, we shouldn't rely on him anymore for
approvals. Mengjiao has been very active with logging and would be a good
addition to the pool of reviewers.

I myself (pohly) have been the de-facto maintainer. Being able to approve
changes from others will be useful and I'll be careful to get a
second approver for my own changes.